move reducers for animation and hotspot to feature
