+++ packages/space-opera/src/components/animation_controls/animation_controls.ts
@@ -22,44 +22,17 @@ import '@polymer/paper-item';
 
 import {customElement, html, internalProperty, query} from 'lit-element';
 
-import {registerStateMutator, State} from '../../redux/space_opera_base.js';
+import {State} from '../../redux/space_opera_base.js';
 import {ConnectedLitElement} from '../connected_lit_element/connected_lit_element.js';
 import {CheckboxElement} from '../shared/checkbox/checkbox.js';
 import {Dropdown} from '../shared/dropdown/dropdown.js';
+
+import {dispatchAnimationName, dispatchAutoplayEnabled, dispatchPlayAnimation} from './reducer.js';
 
 interface AnimationControlsInterface {
   autoplay?: boolean;
   animationName?: string;
 }
-
-// Register state mutators and get corresponding dispatchers.
-
-/** Set auto play enabled or not */
-export const dispatchAutoplayEnabled =
-    registerStateMutator('SET_AUTOPLAY_ENABLED', (state, enabled?: boolean) => {
-      state.config = {...state.config, autoplay: !!enabled};
-    });
-
-/** Set animation name */
-export const dispatchAnimationName = registerStateMutator(
-    'SET_ANIMATION_NAME', (state, animationName?: string) => {
-      // Allow animationName === undefined to unset animationName
-      if (animationName && state.animationNames.indexOf(animationName) === -1) {
-        return;
-      }
-
-      state.config = {
-        ...state.config,
-        animationName,
-      };
-    });
-
-/** Set playAnimation or not */
-export const dispatchPlayAnimation =
-    registerStateMutator('PLAY_ANIMATION', (state, playAnimation?: boolean) => {
-      // No need to copy state - we're always given a new copy.
-      state.playAnimation = !!playAnimation;
-    });
 
 /**
  * Animation controls for gltf and model-viewer.
