+++ packages/space-opera/src/components/hotspot_panel/hotspot_config.ts
