+++ packages/space-opera/src/components/hotspot_panel/reducer.ts
@@ -15,9 +15,10 @@
  *
  */
 
+import {immutableArrayUpdate} from '../../redux/reducer_utils.js';
+import {registerStateMutator, State} from '../../redux/space_opera_base.js';
+
 import {HotspotConfig} from './hotspot_config.js';
-import {immutableArrayUpdate} from './reducer_utils.js';
-import {registerStateMutator, State} from './space_opera_base.js';
 
 let nextHotspotId = 1;
 let hotspotNameSet = new Set();
