+++ packages/space-opera/src/redux/space_opera_base.ts
@@ -19,9 +19,10 @@ import {GltfModel, ModelViewerConfig} from '@google/model-viewer-editing-adapter
 import {ModelViewerElement} from '@google/model-viewer/lib/model-viewer';
 import * as Redux from 'redux';  // from //third_party/javascript/redux:redux_closurized
 
+import {HotspotConfig} from '../components/hotspot_panel/hotspot_config.js';
+
 import {Camera, INITIAL_CAMERA} from './camera_state.js';
 import {getGltfEdits, GltfEdits, INITIAL_GLTF_EDITS} from './gltf_edits.js';
-import {HotspotConfig} from './hotspot_config.js';
 import {INITIAL_ENVIRONMENT_IMAGES} from './initial_environment_images.js';
 import {EnvironmentImage} from './lighting_state.js';
